food101/requirements.txt
## 2) `food101/requirements.txt`

torch
torchvision
open_clip_torch
tqdm
numpy
pandas
pyyaml

